requirements.txt
ultralytics
torch
numpy
opencv-python
pillow
scikit-learn
timm
matplotlib
seaborn